fix notifications and refactor
