+++ app/src/main/java/com/example/todolist/ToDoApp.kt
@@ -2,10 +2,10 @@ package com.example.todolist
 
 import android.app.Application
 import android.content.Context
-import com.example.todolist.data.workers.DataSyncWorker
+import com.example.todolist.data.DataSyncWorker
 import com.example.todolist.di.app.AppComponent
 import com.example.todolist.di.app.DaggerAppComponent
-import com.example.todolist.utils.Constants.SETTINGS
+import com.example.todolist.utils.Constants.SHARED_PREFERENCES_NAME
 import com.example.todolist.utils.setSettingsTheme
 
 class ToDoApp : Application() {
@@ -19,7 +19,7 @@ class ToDoApp : Application() {
             .factory()
             .create(this)
 
-        val settings = getSharedPreferences(SETTINGS, Context.MODE_PRIVATE)
+        val settings = getSharedPreferences(SHARED_PREFERENCES_NAME, Context.MODE_PRIVATE)
         setSettingsTheme(settings)
     }
 }
