+++ app/src/main/java/com/example/todolist/di/activity/ActivityComponent.kt
@@ -1,6 +1,7 @@
 package com.example.todolist.di.activity
 
 import com.example.todolist.di.fragment.EditTaskFragmentComponent
+import com.example.todolist.di.fragment.SettingsFragmentComponent
 import com.example.todolist.di.fragment.ToDoListFragmentComponent
 import com.example.todolist.di.scopes.ActivityScope
 import com.example.todolist.ui.MainActivity
@@ -12,5 +13,5 @@ interface ActivityComponent {
     fun inject(mainActivity: MainActivity)
     fun editTaskFragmentComponent(): EditTaskFragmentComponent
     fun toDoListFragmentComponent(): ToDoListFragmentComponent
-
+    fun settingsFragmentComponent(): SettingsFragmentComponent
 }
