+++ app/src/main/java/com/example/todolist/di/app/AppComponent.kt
@@ -1,15 +1,15 @@
 package com.example.todolist.di.app
 
 import android.content.Context
-import com.example.todolist.data.workers.DataSyncWorker
-import com.example.todolist.data.workers.NotificationBroadcastReceiver
-import com.example.todolist.data.workers.NotificationWorker
+import com.example.todolist.data.DataSyncWorker
+import com.example.todolist.ui.notifications.NotificationWorker
 import com.example.todolist.di.activity.ActivityComponent
 import com.example.todolist.di.scopes.AppScope
+import com.example.todolist.ui.notifications.NotificationBroadcastReceiver
 import dagger.BindsInstance
 import dagger.Component
 
-@Component(modules = [NetworkModule::class, RepositoryModule::class, DbModule::class])
+@Component(modules = [AppModule::class, NetworkModule::class, RepositoryModule::class, DbModule::class])
 @AppScope
 interface AppComponent {
