+++ app/src/main/java/com/example/todolist/ui/Screens/components/ToDoEditField.kt
@@ -1,0 +1,29 @@
+package com.example.todolist.ui.Screens.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+@Composable
+@Preview
+fun EditField(
+    text: String = "",
+    updateTextAction: (text: String) -> Unit = {}
+){
+    Row(
+        modifier = Modifier
+            .padding(top = 8.dp)
+            .fillMaxWidth(),
+        horizontalArrangement = Arrangement.Center
+    ) {
+        EditText(
+            text = text,
+            updateTextAction = updateTextAction
+        )
+    }
+}
