+++ app/src/main/java/com/example/todolist/utils/Constants.kt
@@ -4,13 +4,8 @@ object Constants {
     const val BASEURL = "https://beta.mrdekk.ru/todobackend/"
     const val TOKEN = "unsalesmanlike"
     const val BAD_REQUEST_CODE = 400
-
-    const val NO_IMPORTANCE_POPUP_MENU_BUTTON_ID = 0
-    const val LOW_IMPORTANCE_POPUP_MENU_BUTTON_ID = 1
-    const val HIGH_IMPORTANCE_POPUP_MENU_BUTTON_ID = 2
-
     const val DATABASE_NAME = "todo_database"
-    const val SETTINGS = "settings"
+    const val SHARED_PREFERENCES_NAME = "mySharedPrefs"
     const val THEME = "theme"
     const val LIGHT_THEME = "light"
     const val DARK_THEME = "dark"
