+++ app/src/main/res/layout/task.xml
@@ -1,8 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/rootContainer"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:foreground="?attr/selectableItemBackground">
     <!--        android:buttonTint="@color/checkbox_color_selector"-->
     <com.google.android.material.checkbox.MaterialCheckBox
         android:id="@+id/taskInfoCheckBox"
@@ -19,7 +21,6 @@
         app:layout_constraintStart_toEndOf="@id/taskInfoCheckBox"
         app:layout_constraintEnd_toStartOf="@id/infoTaskButton"
         app:layout_constraintTop_toTopOf="parent"
-        android:foreground="?attr/selectableItemBackground"
         android:paddingTop="12dp"
         android:paddingBottom="16dp">
