--- app/src/main/res/values/color_palette.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-<!--    <color name="back_elevated">#FFFFFF</color>-->
-<!--    <color name="back_primary">#F7F6F2</color>-->
-<!--    <color name="back_secondary">#FFFFFF</color>-->
-<!--    <color name="color_blue">#007AFF</color>-->
-<!--    <color name="color_gray">#8E8E93</color>-->
-<!--    <color name="color_green">#34C759</color>-->
-<!--    <color name="color_red">#FF3B30</color>-->
-<!--    <color name="color_white">#FFFFFF</color>-->
-<!--    <color name="color_gray_light">#D1D1D6</color>-->
-<!--    <color name="label_disable">#26000000</color>-->
-<!--    <color name="label_primary">#000000</color>-->
-<!--    <color name="label_secondary">#99000000</color>-->
-<!--    <color name="label_tertiary">#4D000000</color>-->
-<!--    <color name="support_overlay">#0F000000</color>-->
-<!--    <color name="support_separator">#33000000</color>-->
-</resources>
