+++ app/src/main/res/values/colors_palette.xml
