+++ app/src/main/res/values/typo.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="large_title" parent="Widget.MaterialComponents.TextView">
+        <item name="android:textSize">@dimen/large_title_text_size</item>
+        <item name="android:lineHeight">@dimen/large_title_line_height</item>
+        <item name="android:fontWeight">500</item>
+    </style>
+
+    <style name="title" parent="Widget.MaterialComponents.TextView">
+        <item name="android:textSize">@dimen/title_text_size</item>
+        <item name="android:lineHeight">@dimen/title_line_height</item>
+        <item name="android:fontWeight">500</item>
+    </style>
+
+    <style name="button" parent="Widget.MaterialComponents.TextView">
+        <item name="android:textSize">@dimen/button_text_size</item>
+        <item name="android:lineHeight">@dimen/button_line_height</item>
+        <item name="android:fontWeight">500</item>
+    </style>
+
+    <style name="body" parent="Widget.MaterialComponents.TextView">
+        <item name="android:textSize">@dimen/body_text_size</item>
+        <item name="android:lineHeight">@dimen/body_line_height</item>
+        <item name="android:fontWeight">400</item>
+    </style>
+
+    <style name="subhead" parent="Widget.MaterialComponents.TextView">
+        <item name="android:textSize">@dimen/subhead_text_size</item>
+        <item name="android:lineHeight">@dimen/subhead_line_height</item>
+        <item name="android:fontWeight">400</item>
+    </style>
+</resources>
